Fixed jdyncall runtime on darwin64 and added preliminary optimized code path for functions with 4 args or less.

+++ libraries/Runtime/Comp/src/main/java/jdyncall/PerfTest.java
@@ -44,5 +44,5 @@ public class PerfTest {
         public static native int testASinB(int a, int b);
     }
     public static native int testAddJNI(int a, int b);
-    public static native int testASinB(int a, int b);
+    public static native double testASinB(int a, int b);
 }
